TUC-155 Adding tests in the constant elimination class, for percentiles validation
Some small modifs in other classes.

	modified:   ../../../src/tucucore/computingservice/computingresponse.h
	modified:   ../../../src/tucucore/cyclestatistics.h
	modified:   ../../../src/tucucore/parameterids.h
	modified:   ../../../test/tucucore/drugmodels/buildconstantelimination.h
	modified:   ../../../test/tucucore/drugmodels/test_constanteliminationbolus.h
	modified:   ../../../test/tucucore/pkmodels/constanteliminationbolus.h

+++ src/tucucore/computingservice/computingresponse.h
@@ -202,7 +202,7 @@ public:
 
     void addPercentileData(const std::vector<CycleData> &_data) { m_data.push_back(_data);}
 
-    const CycleData& getData(unsigned int _percentileIndex, unsigned int _cycleIndex) const {
+    const CycleData& getData(size_t _percentileIndex, unsigned int _cycleIndex) const {
         return m_data.at(_percentileIndex).at(_cycleIndex);
     }
